requirements.txt
azure-functions
azure-storage-blob
urllib3
pandas
certifi
twilio